Add self.c and self.h
