+++ src/av/audio.c
@@ -6,6 +6,7 @@
 #include "../tox.h"
 #include "../utox.h"
 #include "../settings.h"
+#include "../self.h"
 
 static void utox_filter_audio_kill(Filter_Audio *filter_audio_handle) {
 #ifdef AUDIO_FILTERING
