+++ src/file_transfers.c
@@ -3,7 +3,9 @@
 #include "friend.h"
 #include "logging_native.h"
 #include "macros.h"
+#include "main.h"
 #include "main_native.h"
+#include "self.h"
 #include "settings.h"
 #include "text.h"
 #include "tox.h"
