+++ src/friend.c
@@ -5,7 +5,9 @@
 #include "filesys.h"
 #include "flist.h"
 #include "logging_native.h"
+#include "main.h"
 #include "main_native.h"
+#include "self.h"
 #include "settings.h"
 #include "text.h"
 #include "tox.h"
