+++ src/messages.c
@@ -10,6 +10,8 @@
 #include "groups.h"
 #include "logging_native.h"
 #include "macros.h"
+#include "main.h"
+#include "self.h"
 #include "settings.h"
 #include "text.h"
 #include "theme.h"
