+++ src/ui/edit.c
@@ -6,6 +6,7 @@
 #include "text.h"
 
 #include "../macros.h"
+#include "../main.h"
 #include "../main_native.h"
 #include "../settings.h"
 #include "../text.h"
